+++ src/PIL/TgaImagePlugin.py
@@ -20,7 +20,6 @@
 import warnings
 
 from . import Image, ImageFile, ImagePalette
-from ._binary import i8
 from ._binary import i16le as i16
 from ._binary import o8
 from ._binary import o16le as o16
@@ -56,14 +55,14 @@ class TgaImageFile(ImageFile.ImageFile):
         # process header
         s = self.fp.read(18)
 
-        id_len = i8(s[0])
+        id_len = s[0]
 
-        colormaptype = i8(s[1])
-        imagetype = i8(s[2])
+        colormaptype = s[1]
+        imagetype = s[2]
 
-        depth = i8(s[16])
+        depth = s[16]
 
-        flags = i8(s[17])
+        flags = s[17]
 
         self._size = i16(s[12:]), i16(s[14:])
